1.Api/requirements.txt
Django==4.0
djangorestframework==3.13.1
psycopg2==2.9.2
drf-yasg==1.20.0
django-cors-headers==3.10.1